added more logic of the system

+++ src/MulticastThreads/MulticastChanelData.java
@@ -54,7 +54,7 @@ public class MulticastChanelData extends MulticastChanel {
                 switch (message.getMsgType()){
                     case PUTCHUNK:
                         // Checking if chunk is already saved by this peer.
-                        if(VolatileDatabase.check_has_chunk(message.getFileId()), message.getChunkNo())
+                        if(VolatileDatabase.check_has_chunk(message.getFileId(), Short.valueOf(message.getChunkNo())))
                             break;
 
                         Peer.threadPool.schedule(new PutChunkHandle(message),delay, TimeUnit.MILLISECONDS);
